refactor: JS to TS (non-strict); resolve any's later
