+++ static_report/src/components/ComparisonReportList.tsx
