+++ static_report/src/components/Loading.tsx
