+++ static_report/src/components/Main.tsx
