+++ static_report/src/components/NotFound.tsx
