+++ static_report/src/components/SingleReport.tsx
