+++ static_report/src/hooks/useDocumentTitle.tsx
