+++ static_report/src/hooks/useHashLcocation.tsx
