+++ static_report/src/utils/theme.ts
